+++ components/MenuNew/MenuSeparator.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './styles.css';
+
+const MenuSeparator = () => {
+    return (
+        <div className={styles.menuSeparator} />
+    );
+};
+
+export default MenuSeparator;